This Jupyter notebook facilitated my global analysis of experimental transient absorption spectra.
A colleague shared this Jupyter notebook with me. While it functioned under Python 2, it encountered issues when attempted in Python 3 due to a range of factors. After implementing essential modifications in the code and related modules, I tackled discrepancies in syntax, library updates, and compatibility matters. As a result, the notebook is now fully operational with Python 3.